Updated Angular

+++ Angular/ITP/src/app/admin/admin.component.ts
@@ -23,7 +23,8 @@ export class AdminComponent implements OnInit {
   reactiveForm: FormGroup;
   employee:Employee;
   type:boolean=true;
-
+  empType:boolean;
+  title:string;
   constructor(private adminService: AdminService, private homeService: HomeService,private router:Router,private fb: FormBuilder) {
     this.reactiveForm = this.fb.group({
     'empName'  : [null,Validators.compose([Validators.required,Validators.minLength(1),Validators.pattern('[A-Za-z]+ *[A-Za-z ]+$')])],
@@ -31,6 +32,7 @@ export class AdminComponent implements OnInit {
     'empPassword'  : [null,Validators.compose([Validators.required,Validators.minLength(8)])],
     'empContact'  : [null,Validators.compose([Validators.required,Validators.minLength(10),Validators.maxLength(10),Validators.pattern('^[7-9][0-9]{9}$')])],
     })
+    
   }
 
   ngOnInit() {
@@ -39,18 +41,16 @@ export class AdminComponent implements OnInit {
     this.username = name[0];
     console.log( JSON.parse(localStorage.getItem('authenticationObject')));
     console.log(JSON.parse(localStorage.getItem('authenticationObject')));
-
-   this.getManagers();
+    this.getManagers();
     this.getAllEmployees();
+  
   }
   logOut(): void {
     this.adminService.logOutUser().then(response => {
       console.log(this.authentication.statusCode);
       this.authentication = response;
-      console.log(response);
-      localStorage.clear();
-      localStorage.removeItem('authenticationObject');
       if (this.authentication.statusCode == 1){
+        localStorage.removeItem('authenticationObject');
         localStorage.clear();
         this.router.navigate(['/']);
       }
@@ -64,15 +64,19 @@ export class AdminComponent implements OnInit {
     return this.employees;
   }
   getAllEmployees():void{
-    this.adminService.getEmployees().then( response =>{ this.employees = response  } );
+    this.adminService.getEmployees().then( response =>{ this.employees = response ;
+    
+    } );
   }
   editEmployeeDetail(employee:Employee):void{
-   this.type = false;
+  this.type = false;
   this.getEmployeeDetail(employee);
   }
   deleteEmployee(employee:Employee):void{
     console.log(employee);
+    if ( confirm ('Are you sure you want to delete the employee : ' + employee.name )) {
     this.adminService.deleteEmployee(employee).then(response => this.authentication = response);
+    }
   }
 
   getDetails(employee:Employee):void{
@@ -80,16 +84,39 @@ export class AdminComponent implements OnInit {
     this.getEmployeeDetail(employee);
   }
   addManager(employee:Employee):void{
-    this.adminService.addManager(employee).then(response=> this.authentication = response);
-    location.reload(true);
+    this.adminService.addManager(employee).then(response=>
+      { 
+        this.authentication = response;
+        location.reload(true);
+      });
   }
-
-
-
-
   getEmployeeDetail(employee:Employee):void{
     console.log(employee);
     console.log("getting type" + this.type);
-this.router.navigate(['/employeeDetail', employee.login.username,this.type ]);
+    this.router.navigate(['/employeeDetail', employee.login.username,this.type ]);
+  }
+  registerEmployee(name:string,username:string,password:string,contactNo:number){
+    let input = username.split("@");
+    console.log(username);
+   console.log(input[0]);
+    let domain = input[1];
+    console.log("in domain" + domain);
+    this.homeService.signUp(name,password,username,contactNo,domain).then(
+      response =>{ this.authentication = response;
+                   if(this.authentication.statusCode ==1){
+                     this.router.navigate(['adminDashboard']);
+                     location.reload(true);
+                   }                     });
+  }
+
+  checkDomain(email:string){
+    let domain=this.autheticatedHeader.username.split('@')[1];
+    let userDomain=email.split('@')[1];
+    if(domain==userDomain){
+      this.title='';
+    }
+    else{
+      this.title='Specify correct domain name';
+    }
   }
 }
